feat: a lot of things that I forgot to commit
- I don't remember what I have added

Took 42 minutes
