+++ lib/logic/canvas_service.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 class CanvasState {
