+++ lib/logic/component_index_service.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/foundation.dart';
 
 base class ComponentsIndexService {
