+++ lib/logic/global_state_service.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/foundation.dart';
 
 class GlobalState {
