+++ lib/logic/tool_service.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:flutter/foundation.dart';
 
 class Tool {
