+++ lib/main.dart
@@ -12,6 +12,9 @@ import 'package:editicert/logic/tool_service.dart';
 import 'package:editicert/utils.dart';
 import 'package:editicert/widgets/controller_widget.dart';
 import 'package:editicert/widgets/creator_widget.dart';
+import 'package:editicert/widgets/left_sidebar.dart';
+import 'package:editicert/widgets/right_sidebar.dart';
+import 'package:editicert/widgets/top_bar.dart';
 import 'package:flex_color_picker/flex_color_picker.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -387,18 +390,6 @@ class _HomePageState extends State<HomePage> with GetItStateMixin {
               // Canvas
               child: Stack(
                 children: [
-                  // unselect all
-                  GestureDetector(
-                    onTap: () {
-                      selectedNotifier.clear();
-                      hoveredNotifier.clear();
-                    },
-                    child: Container(
-                      width: mqSize.width,
-                      height: mqSize.height,
-                      color: Colors.transparent,
-                    ),
-                  ),
                   // Background/canvas
                   ValueListenableBuilder(
                     valueListenable: transformationController,
@@ -410,6 +401,8 @@ class _HomePageState extends State<HomePage> with GetItStateMixin {
                           onTap: () {
                             selectedNotifier.clear();
                             hoveredNotifier.clear();
+                            isShowColorPicker = false;
+                            setState(() {});
                           },
                           child: Container(
                             width: mqSize.width,
@@ -591,396 +584,12 @@ class _HomePageState extends State<HomePage> with GetItStateMixin {
             Column(
               children: [
                 /// Top bar
-                Container(
-                  height: topbarHeight,
-                  decoration: BoxDecoration(
-                    color: colorScheme.surface,
-                    border: Border(
-                      bottom: BorderSide(
-                        color: Colors.grey.withOpacity(.375),
-                        width: 1,
-                      ),
-                    ),
-                  ),
-                  padding: EdgeInsets.only(
-                    left: 8 +
-                        (Platform.isMacOS &&
-                                !globalStateNotifier.state.value.states
-                                    .contains(GlobalStates.fullscreen)
-                            ? 80
-                            : 0),
-                    right: 8,
-                  ),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      ...[
-                        (
-                          text: 'Move',
-                          shortcut: 'V',
-                          tool: ToolType.move,
-                          onTap: () {
-                            toolNotifier.setMove();
-                          },
-                          icon: Transform.rotate(
-                            angle: -pi / 5,
-                            alignment: const Alignment(-0.2, 0.3),
-                            child: const Icon(
-                              Icons.navigation_outlined,
-                              size: 18,
-                            ),
-                          )
-                        ),
-                        (
-                          text: 'Frame',
-                          shortcut: 'F',
-                          tool: ToolType.frame,
-                          onTap: () {
-                            toolNotifier.setFrame();
-                          },
-                          icon: const Icon(
-                            CupertinoIcons.grid,
-                            size: 18,
-                          )
-                        ),
-                        (
-                          text: 'Rectangle',
-                          shortcut: 'R',
-                          tool: ToolType.rectangle,
-                          onTap: () {
-                            toolNotifier.setRectangle();
-                          },
-                          icon: const Icon(
-                            CupertinoIcons.square,
-                            size: 18,
-                          )
-                        ),
-                        (
-                          text: 'Hand',
-                          shortcut: 'H',
-                          tool: ToolType.hand,
-                          onTap: () {
-                            toolNotifier.setHand();
-                          },
-                          icon: const Icon(
-                            CupertinoIcons.hand_raised,
-                            size: 18,
-                          )
-                        ),
-                        (
-                          text: 'Text',
-                          shortcut: 'T',
-                          tool: ToolType.text,
-                          onTap: () {
-                            toolNotifier.setText();
-                          },
-                          icon: const Icon(
-                            CupertinoIcons.textbox,
-                            size: 18,
-                          )
-                        ),
-                      ].map(
-                        (e) => Tooltip(
-                          richMessage: TextSpan(
-                            children: [
-                              TextSpan(
-                                text: '${e.text} ',
-                              ),
-                              TextSpan(
-                                text: ' ${e.shortcut}',
-                                style: TextStyle(
-                                  fontWeight: FontWeight.w900,
-                                  color: Colors.grey.shade600,
-                                ),
-                              ),
-                            ],
-                          ),
-                          child: Card(
-                            margin: EdgeInsets.zero,
-                            color: tool == e.tool
-                                ? colorScheme.onSurface.withOpacity(.125)
-                                : colorScheme.surface,
-                            elevation: 0,
-                            clipBehavior: Clip.hardEdge,
-                            shape: const RoundedRectangleBorder(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(4)),
-                            ),
-                            child: InkWell(
-                              onTap: e.onTap,
-                              child: Padding(
-                                padding: const EdgeInsets.all(17.0),
-                                child: e.icon,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      TextButton.icon(
-                        style: ButtonStyle(
-                          shape: MaterialStateProperty.all(
-                            ContinuousRectangleBorder(
-                              borderRadius: BorderRadius.circular(16),
-                            ),
-                          ),
-                        ),
-                        onPressed: () {
-                          transformationController.value =
-                              Matrix4.identity().scaled(
-                            transformationController.value.getMaxScaleOnAxis(),
-                            transformationController.value.getMaxScaleOnAxis(),
-                          );
-                        },
-                        icon: const Icon(
-                          Icons.navigation_rounded,
-                          size: 16,
-                        ),
-                        label: const Text('Recenter'),
-                      ),
-                      const SizedBox(width: 8),
-                      ValueListenableBuilder(
-                        valueListenable: transformationController,
-                        builder: (context, value, child) {
-                          return Text(
-                            '${(value.getMaxScaleOnAxis() * 100).truncate()}%',
-                          );
-                        },
-                      ),
-                    ],
-                  ),
-                ),
+                TopBar(),
                 Expanded(
                   child: Row(
                     children: [
                       /// Left Sidebar
-                      Container(
-                        width: sidebarWidth,
-                        color: colorScheme.surface,
-                        child: ListView(
-                          children: components
-                              .mapIndexed(
-                                (i, e) => Container(
-                                  padding: const EdgeInsets.only(right: 8),
-                                  height: 32,
-                                  decoration: BoxDecoration(
-                                    color: selected.contains(i)
-                                        ? colorScheme.surfaceVariant
-                                        : null,
-                                    border: Border.all(
-                                      strokeAlign: BorderSide.strokeAlignInside,
-                                      color: !hovered.contains(i) ||
-                                              selected.contains(i)
-                                          ? Colors.transparent
-                                          : Colors.blueAccent.withOpacity(.5),
-                                    ),
-                                  ),
-                                  // color: Colors.transparent,
-                                  child: MouseRegion(
-                                    onEnter: (event) => hoveredNotifier.add(i),
-                                    onExit: (event) =>
-                                        hoveredNotifier.remove(i),
-                                    child: InkWell(
-                                      onTap: () {
-                                        selectedNotifier
-                                          ..clear()
-                                          ..add(i);
-                                      },
-                                      child: Row(
-                                        children: [
-                                          const Padding(
-                                            padding: EdgeInsets.symmetric(
-                                              horizontal: 8.0,
-                                            ),
-                                            child: Icon(
-                                              Icons.rectangle_outlined,
-                                              size: 12,
-                                            ),
-                                          ),
-                                          Expanded(
-                                            child: Text(
-                                              e.name.replaceAll('\n', ' '),
-                                              maxLines: 1,
-                                              overflow: TextOverflow.ellipsis,
-                                              style: Theme.of(context)
-                                                  .textTheme
-                                                  .bodySmall,
-                                            ),
-                                          ),
-                                          if (hovered.contains(i))
-                                            SizedBox(
-                                              width: 18 * 2,
-                                              child: Row(
-                                                mainAxisAlignment:
-                                                    MainAxisAlignment
-                                                        .spaceBetween,
-                                                children: [
-                                                  IconButton(
-                                                    hoverColor:
-                                                        Colors.transparent,
-                                                    splashColor:
-                                                        Colors.transparent,
-                                                    focusColor:
-                                                        Colors.transparent,
-                                                    highlightColor:
-                                                        Colors.transparent,
-                                                    padding: EdgeInsets.zero,
-                                                    constraints:
-                                                        BoxConstraints.tight(
-                                                      Size(
-                                                        e.locked ? 14 : 18,
-                                                        double.infinity,
-                                                      ),
-                                                    ),
-                                                    onPressed: () =>
-                                                        componentsNotifier
-                                                            .replace(
-                                                      i,
-                                                      locked: !e.locked,
-                                                    ),
-                                                    icon: Icon(
-                                                      e.locked
-                                                          ? CupertinoIcons
-                                                              .lock_fill
-                                                          : CupertinoIcons
-                                                              .lock_open_fill,
-                                                      size: 14,
-                                                    ),
-                                                  ),
-                                                  IconButton(
-                                                    hoverColor:
-                                                        Colors.transparent,
-                                                    splashColor:
-                                                        Colors.transparent,
-                                                    focusColor:
-                                                        Colors.transparent,
-                                                    highlightColor:
-                                                        Colors.transparent,
-                                                    padding: EdgeInsets.zero,
-                                                    constraints:
-                                                        BoxConstraints.tight(
-                                                      const Size(
-                                                        18,
-                                                        double.infinity,
-                                                      ),
-                                                    ),
-                                                    onPressed: () =>
-                                                        componentsNotifier
-                                                            .replace(
-                                                      i,
-                                                      hidden: !e.hidden,
-                                                    ),
-                                                    icon: Icon(
-                                                      e.hidden
-                                                          ? CupertinoIcons
-                                                              .eye_slash
-                                                          : CupertinoIcons.eye,
-                                                      size: 14,
-                                                    ),
-                                                  ),
-                                                ],
-                                              ),
-                                            ),
-                                        ],
-                                      ),
-                                    ),
-                                    // child: ListTile(
-                                    //   textColor: e.hidden ? Colors.grey : null,
-                                    //   iconColor: e.hidden ? Colors.grey : null,
-                                    //   onTap: () {
-                                    // selected
-                                    //       ..clear()
-                                    //       ..add(i);
-                                    //   },
-                                    //   hoverColor: Colors.transparent,
-                                    //   // selected: selected.contains(i),
-                                    //   horizontalTitleGap: 4,
-                                    //   title: Text(
-                                    //     e.name,
-                                    //     style: const TextStyle(fontSize: 12),
-                                    //   ),
-                                    //   leading: const Icon(
-                                    //     CupertinoIcons.square,
-                                    //     size: 16,
-                                    //   ),
-                                    //   contentPadding: const EdgeInsets.only(
-                                    //       left: 4, right: 6),
-                                    //   trailing: hovered.contains(i)
-                                    //       ? SizedBox(
-                                    //           width: 18 * 2,
-                                    //           child: Row(
-                                    //             mainAxisAlignment:
-                                    //                 MainAxisAlignment
-                                    //                     .spaceBetween,
-                                    //             children: [
-                                    //               IconButton(
-                                    //                 hoverColor:
-                                    //                     Colors.transparent,
-                                    //                 splashColor:
-                                    //                     Colors.transparent,
-                                    //                 focusColor:
-                                    //                     Colors.transparent,
-                                    //                 highlightColor:
-                                    //                     Colors.transparent,
-                                    //                 padding: EdgeInsets.zero,
-                                    //                 constraints: BoxConstraints
-                                    //                     .tight(Size(
-                                    //                         e.locked ? 14 : 18,
-                                    //                         double.infinity)),
-                                    //                 onPressed: () =>
-                                    //                     .read(componentsProvider
-                                    //                         .notifier)
-                                    //                     .replace(i,
-                                    //                         locked: !e.locked),
-                                    //                 icon: Icon(
-                                    //                   e.locked
-                                    //                       ? CupertinoIcons
-                                    //                           .lock_fill
-                                    //                       : CupertinoIcons
-                                    //                           .lock_open_fill,
-                                    //                   size: 14,
-                                    //                 ),
-                                    //               ),
-                                    //               IconButton(
-                                    //                 hoverColor:
-                                    //                     Colors.transparent,
-                                    //                 splashColor:
-                                    //                     Colors.transparent,
-                                    //                 focusColor:
-                                    //                     Colors.transparent,
-                                    //                 highlightColor:
-                                    //                     Colors.transparent,
-                                    //                 padding: EdgeInsets.zero,
-                                    //                 constraints:
-                                    //                     BoxConstraints.tight(
-                                    //                         const Size(
-                                    //                             18,
-                                    //                             double
-                                    //                                 .infinity)),
-                                    //                 onPressed: () =>
-                                    //                     .read(componentsProvider
-                                    //                         .notifier)
-                                    //                     .replace(i,
-                                    //                         hidden: !e.hidden),
-                                    //                 icon: Icon(
-                                    //                   e.hidden
-                                    //                       ? CupertinoIcons
-                                    //                           .eye_slash
-                                    //                       : CupertinoIcons.eye,
-                                    //                   size: 14,
-                                    //                 ),
-                                    //               ),
-                                    //             ],
-                                    //           ),
-                                    //         )
-                                    //       : null,
-                                    // ),
-                                  ),
-                                ),
-                              )
-                              .toList(),
-                        ),
-                      ),
+                      LeftSidebar(),
                       const Expanded(
                         child: SizedBox.shrink(),
                       ),
@@ -1363,399 +972,6 @@ class _HomePageState extends State<HomePage> with GetItStateMixin {
             child: child,
           ),
         ),
-      ),
-    );
-  }
-}
-
-class RightSidebar extends StatefulWidget with GetItStatefulWidgetMixin {
-  RightSidebar({
-    super.key,
-    required this.toggleColorPicker,
-  });
-
-  final VoidCallback toggleColorPicker;
-
-  @override
-  State<RightSidebar> createState() => _RightSidebarState();
-}
-
-class _RightSidebarState extends State<RightSidebar> with GetItStateMixin {
-  late final TextEditingController backgroundColorController;
-  late final TextEditingController backgroundWidthController;
-  late final TextEditingController backgroundHeightController;
-
-  @override
-  void initState() {
-    backgroundColorController = TextEditingController(
-      text: canvasStateNotifier.state.value.color.value
-          .toRadixString(16)
-          .toUpperCase()
-          .substring(2),
-    );
-    backgroundWidthController = TextEditingController(
-      text: canvasStateNotifier.state.value.size.width.toString(),
-    );
-    backgroundHeightController = TextEditingController(
-      text: canvasStateNotifier.state.value.size.height.toString(),
-    );
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final selected = watchX((Selected selected) => selected.state);
-    final canvasStateProvider =
-        watchX((CanvasState canvasState) => canvasState.state);
-    final component = selected.firstOrNull == null
-        ? null
-        : componentsNotifier.state.value[selected.first].component;
-    final textTheme = Theme.of(context).textTheme;
-
-    final controls = component == null
-        ? null
-        : [
-            (
-              children: [
-                (
-                  prefix: Text(
-                    'X',
-                    style: textTheme.bodySmall,
-                    textAlign: TextAlign.center,
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text: component.pos.dx.toStringAsFixed(1),
-                  ),
-                ),
-                (
-                  prefix: Text(
-                    'Y',
-                    style: textTheme.bodySmall,
-                    textAlign: TextAlign.center,
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text: component.pos.dy.toStringAsFixed(1),
-                  ),
-                ),
-              ]
-            ),
-            (
-              children: [
-                (
-                  prefix: Text(
-                    'W',
-                    style: textTheme.bodySmall,
-                    textAlign: TextAlign.center,
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text: component.size.width.toStringAsFixed(1),
-                  ),
-                ),
-                (
-                  prefix: Text(
-                    'H',
-                    style: textTheme.bodySmall,
-                    textAlign: TextAlign.center,
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text: component.size.height.toStringAsFixed(1),
-                  ),
-                ),
-              ]
-            ),
-            (
-              children: [
-                (
-                  prefix: Transform.translate(
-                    offset: const Offset(0, -1),
-                    child: const Icon(
-                      size: 14,
-                      CupertinoIcons.rotate_right,
-                    ),
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text:
-                        '${((component.angle % (pi * 2)) / pi * 180).toStringAsFixed(1)}°',
-                  ),
-                ),
-                (
-                  prefix: Transform.translate(
-                    offset: const Offset(0, 1),
-                    child: const Icon(
-                      Icons.rounded_corner_rounded,
-                      size: 12,
-                    ),
-                  ),
-                  keyboardType: TextInputType.number,
-                  controller: TextEditingController(
-                    text: '0',
-                  ),
-                ),
-              ]
-            ),
-          ];
-
-    final backgroundControl = SizedBox(
-      height: 16,
-      child: Row(
-        children: [
-          SizedBox(
-            width: textFieldWidth,
-            child: Row(
-              children: [
-                Container(
-                  width: 16,
-                  height: 16,
-                  margin: const EdgeInsets.only(right: 6),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(2),
-                    color: canvasStateProvider.color,
-                  ),
-                  child: InkWell(
-                    onTap: widget.toggleColorPicker,
-                    child: const SizedBox.shrink(),
-                  ),
-                ),
-                Expanded(
-                  child: TextField(
-                    maxLength: 6,
-                    onChanged: (value) => canvasStateNotifier.update(
-                      backgroundColor: value.toColor,
-                    ),
-                    controller: backgroundColorController,
-                    style: textTheme.bodySmall,
-                    decoration: const InputDecoration(
-                      hintText: '',
-                      counter: SizedBox.shrink(),
-                      border: InputBorder.none,
-                      contentPadding: EdgeInsets.only(top: 32),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 4.0),
-            child: Text(
-              '${(canvasStateProvider.color.opacity * 100).truncate()}%',
-              style: textTheme.bodySmall,
-            ),
-          ),
-          Expanded(
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                IconButton(
-                  hoverColor: Colors.transparent,
-                  splashColor: Colors.transparent,
-                  focusColor: Colors.transparent,
-                  highlightColor: Colors.transparent,
-                  padding: EdgeInsets.zero,
-                  constraints:
-                      BoxConstraints.tight(const Size(18, double.infinity)),
-                  onPressed: () => canvasStateNotifier.update(
-                    backgroundHidden: !canvasStateProvider.hidden,
-                  ),
-                  icon: Icon(
-                    canvasStateProvider.hidden
-                        ? CupertinoIcons.eye_slash
-                        : CupertinoIcons.eye,
-                    size: 14,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-
-    final canvasSizeControl = [
-      (
-        children: [
-          (
-            prefix: Text(
-              'W',
-              style: textTheme.bodySmall,
-              textAlign: TextAlign.center,
-            ),
-            keyboardType: TextInputType.number,
-            controller: backgroundWidthController,
-            onChanged: (text) {
-              print('test');
-              print(canvasStateProvider.size);
-              canvasStateNotifier.update(
-                  backgroundSize: Size(
-                double.parse(text),
-                canvasStateProvider.size.height,
-              ));
-              print(canvasStateProvider.size);
-            }
-          ),
-          (
-            prefix: Text(
-              'H',
-              style: textTheme.bodySmall,
-              textAlign: TextAlign.center,
-            ),
-            keyboardType: TextInputType.number,
-            controller: backgroundHeightController,
-            onChanged: (text) => canvasStateNotifier.update(
-                    backgroundSize: Size(
-                  canvasStateProvider.size.width,
-                  double.parse(text),
-                ))
-          ),
-        ]
-      ),
-    ].map(
-      (e) => SizedBox(
-        height: 32,
-        child: Row(
-          children: [
-            ...e.children.map(
-              (e) => SizedBox(
-                width: textFieldWidth,
-                child: Row(
-                  children: [
-                    // w: 24
-                    Container(
-                      width: 16,
-                      height: 16,
-                      margin: const EdgeInsets.only(right: 6),
-                      child: e.prefix,
-                    ),
-                    // w: 72
-                    Expanded(
-                      child: TextField(
-                        onChanged: e.onChanged,
-                        cursorHeight: 12,
-                        style: textTheme.bodySmall,
-                        decoration: const InputDecoration.collapsed(
-                          hintText: '',
-                        ),
-                        keyboardType: e.keyboardType,
-                        controller: e.controller,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-
-    return Container(
-      width: sidebarWidth,
-      decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.surface,
-      ),
-      child: ListView(
-        children: [
-          ...[
-            (
-              title: Text(
-                'Background',
-                style: textTheme.labelMedium,
-              ),
-              contents: [
-                backgroundControl,
-                const SizedBox(
-                  height: 8,
-                ),
-                ...canvasSizeControl,
-              ]
-            ),
-            if (component != null)
-              (
-                title: Text(
-                  'Component',
-                  style: textTheme.labelMedium,
-                ),
-                contents: [
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      ...controls!.mapIndexed(
-                        (i, e) => Padding(
-                          padding: EdgeInsets.only(
-                              bottom: i == controls.length - 1 ? 8 : 16.0),
-                          child: SizedBox(
-                            height: 16,
-                            child: Row(
-                              children: [
-                                ...e.children.map(
-                                  (e) => SizedBox(
-                                    width: textFieldWidth,
-                                    child: Row(
-                                      children: [
-                                        // w: 24
-                                        Container(
-                                          width: 16,
-                                          height: 16,
-                                          margin:
-                                              const EdgeInsets.only(right: 6),
-                                          child: e.prefix,
-                                        ),
-                                        // w: 72
-                                        Expanded(
-                                          child: TextField(
-                                            cursorHeight: 12,
-                                            style: textTheme.bodySmall,
-                                            decoration:
-                                                const InputDecoration.collapsed(
-                                              hintText: '',
-                                            ),
-                                            keyboardType: e.keyboardType,
-                                            controller: e.controller,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ]
-              ),
-          ].map(
-            (e) {
-              final title = e.title;
-              return Container(
-                padding:
-                    const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
-                decoration: BoxDecoration(
-                  border: Border(
-                    bottom: BorderSide(
-                      color: Colors.grey.withOpacity(.5),
-                      width: 1,
-                    ),
-                  ),
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    if (title != null) SizedBox(height: 32, child: title),
-                    ...e.contents.map((e) => e),
-                  ],
-                ),
-              );
-            },
-          ),
-        ],
       ),
     );
   }
