+++ lib/widgets/controller_widget.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:math';
 
 import 'package:editicert/logic/component_index_service.dart';
