+++ pubspec.yaml
@@ -4,7 +4,7 @@ description: Certificate Editor & Batch Exporter
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ">=3.1.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dependencies:
   collection: ^1.17.1
   transparent_pointer: ^1.0.0
   flutter_launcher_icons: ^0.13.1
-  window_manager: ^0.3.4
+  window_manager: ^0.3.6
   macos_window_utils: ^1.1.3
   touch_bar: ^0.3.0-alpha.1
   get_it: ^7.6.0
@@ -30,25 +30,25 @@ dev_dependencies:
 
   build_runner: ^2.4.4
   flutter_lints: ^2.0.0
-  dart_code_metrics: ^5.7.4
+  dart_code_metrics_presets: ^2.3.0
 
 flutter:
   uses-material-design: true
 
 flutter_launcher_icons:
-#  android: "launcher_icon"
-#  ios: true
-#  image_path: "assets/icon/icon.png"
-#  min_sdk_android: 21 # android min sdk min:16, default 21
-#  web:
-#    generate: true
-#    image_path: "path/to/image.png"
-#    background_color: "#hexcode"
-#    theme_color: "#hexcode"
-#  windows:
-#    generate: true
-#    image_path: "path/to/image.png"
-#    icon_size: 48 # min:48, max:256, default: 48
+  #  android: "launcher_icon"
+  #  ios: true
+  #  image_path: "assets/icon/icon.png"
+  #  min_sdk_android: 21 # android min sdk min:16, default 21
+  #  web:
+  #    generate: true
+  #    image_path: "path/to/image.png"
+  #    background_color: "#hexcode"
+  #    theme_color: "#hexcode"
+  #  windows:
+  #    generate: true
+  #    image_path: "path/to/image.png"
+  #    icon_size: 48 # min:48, max:256, default: 48
   macos:
     generate: true
     image_path: "assets/app_icon_1024.png"
